handle add product , product specification

+++ src/main/java/com/shopelec/backend/service/brand/BrandServiceImpl.java
@@ -7,6 +7,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@@ -35,7 +37,22 @@ public class BrandServiceImpl implements BrandService{
     }
 
     @Override
-    public void delete(Long id) {
-        brandRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if(brandRepository.existsById(id)) {
+            brandRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public boolean existByName(String name) {
+        return brandRepository.existsByName(name);
+    }
+
+    @Override
+    public List<Brand> getAllBrand() {
+        return brandRepository.findAll();
     }
 }
